webvirtbackend/requirements/production.txt
Django==4.1.3
celery==5.2.7
passlib==1.7.4
gunicorn==20.1.0
mysqlclient==2.1.1
crispy-tailwind==0.5.0
django-crispy-forms==2.0
django_celery_results==2.4.0
djangorestframework==3.14.0